+++ CotEditor/Sources/Console.swift
@@ -102,7 +102,7 @@ final class ConsoleViewController: NSViewController {
     
     private let dateFormatter: DateFormatter = {
         let formatter = DateFormatter()
-        formatter.dateFormat = "YYYY-MM-dd HH:mm:ss"
+        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
         return formatter
     }()