+++ CotEditor/Sources/NSTextView+MultiCursor.swift
@@ -241,11 +241,11 @@ extension MultiCursorEditing where Self: NSTextView {
     /// Move all cursors and expand selection with the same rule.
     ///
     /// - Parameters:
+    ///   - forward: `true` if the cursor should move forward, otherwise `false`.
     ///   - affinity: The selection affinity for the movement.
-    ///   - block: The block that describes the rule how to change the selection.
-    ///   - range: The range of each insertion.
-    ///   - origin: The character index where the selection initially started.
-    func moveCursorsAndModifySelection(affinity: NSSelectionAffinity, using block: (_ range: NSRange, _ origin: Int?) -> (cursor: Int, origin: Int)) {
+    ///   - block: The block that describes the rule how to move the cursor.
+    ///   - cursor: The character index of the cursor to move.
+    func moveCursorsAndModifySelection(forward: Bool, affinity: NSSelectionAffinity, using block: (_ cursor: Int) -> Int) {
         
         var origins = self.selectionOrigins
         var newOrigins: [Int] = []
@@ -257,11 +257,20 @@ extension MultiCursorEditing where Self: NSTextView {
                 origin = nil
             }
             
-            let bounds = block(range, origin)
+            let (cursor, newOrigin): (Int, Int) = {
+                switch (forward, origin) {
+                case (false, range.lowerBound): return (range.upperBound, range.lowerBound)
+                case (false, _):                return (range.lowerBound, range.upperBound)
+                case (true, range.upperBound):  return (range.lowerBound, range.upperBound)
+                case (true, _):                 return (range.upperBound, range.lowerBound)
+                }
+            }()
             
-            newOrigins.append(origin ?? bounds.origin)
+            let newCursor = block(cursor)
             
-            return (bounds.cursor <= bounds.origin) ? NSRange(bounds.cursor..<bounds.origin) : NSRange(bounds.origin..<bounds.cursor)
+            newOrigins.append(origin ?? newOrigin)
+            
+            return (newCursor <= newOrigin) ? NSRange(newCursor..<newOrigin) : NSRange(newOrigin..<newCursor)
         }
         
         guard let set = self.prepareForSelectionUpdate(ranges) else { return assertionFailure() }
@@ -298,58 +307,78 @@ extension MultiCursorEditing where Self: NSTextView {
             else { assertionFailure(); return }
         
         let glyphRanges = self.insertionRanges.map { layoutManager.glyphRange(forCharacterRange: $0, actualCharacterRange: nil) }
-        let wholeRange = NSRange(glyphRanges.first!.lowerBound..<glyphRanges.last!.upperBound)
-        let lineFragmentUsedRects = layoutManager.lineFragmentUsedRects(in: wholeRange)
+        var effectiveGlyphRange: NSRange = .notFound
+        let lineFragmentUsedRects = layoutManager.lineFragmentUsedRects(inSelectedGlyphRanges: glyphRanges, effectiveRange: &effectiveGlyphRange)
         
-        // new visual line to append
-        var newLineRect: NSRect = {
+        // abort when one of the cusors already reached to the edge
+        guard
+            !(affinity == .downstream && effectiveGlyphRange.lowerBound == 0),
+            !(affinity == .upstream && (
+                (layoutManager.extraLineFragmentTextContainer == nil && effectiveGlyphRange.upperBound == layoutManager.numberOfGlyphs) ||
+                (layoutManager.extraLineFragmentTextContainer != nil && glyphRanges.last?.lowerBound == layoutManager.numberOfGlyphs)))
+            else { return }
+        
+        // get new visual line to append
+        // -> Use line fragment to allow placing insertion points even when the line is shorter than the origin insertion columns.
+        let newLineRect: CGRect = {
             switch affinity {
             case .downstream:
-                let rect = lineFragmentUsedRects.first!
-                return rect.offsetBy(dx: 0, dy: -rect.height)
-            case.upstream:
-                let rect = lineFragmentUsedRects.last!
-                return rect.offsetBy(dx: 0, dy: rect.height)
+                return layoutManager.lineFragmentRect(forGlyphAt: effectiveGlyphRange.lowerBound - 1, effectiveRange: nil, withoutAdditionalLayout: true)
+            case .upstream where effectiveGlyphRange.upperBound < layoutManager.numberOfGlyphs:
+                return layoutManager.lineFragmentRect(forGlyphAt: effectiveGlyphRange.upperBound, effectiveRange: nil, withoutAdditionalLayout: true)
+            case .upstream:
+                return layoutManager.extraLineFragmentRect
             }
         }()
         
-        let baseIndex = (affinity == .downstream) ? glyphRanges.last!.lowerBound : glyphRanges.first!.upperBound
-        
         // get base selection rects in the origin line
-        // -> At the same time, expand newLineRect to the entire line fragment
-        //    to allow placing insertion points even when the line is shorter than the origin insertion columns.
+        let baseIndex = (affinity == .downstream) ? glyphRanges.last!.lowerBound : glyphRanges.first!.upperBound
         var baseLineRange: NSRange = .notFound
-        newLineRect.size.width = layoutManager.lineFragmentRect(forGlyphAt: baseIndex, effectiveRange: &baseLineRange, withoutAdditionalLayout: true).width
+        layoutManager.lineFragmentRect(forGlyphAt: min(baseIndex, layoutManager.numberOfGlyphs - 1), effectiveRange: &baseLineRange, withoutAdditionalLayout: true)
         let rowBounds = glyphRanges
             .filter { baseLineRange.touches($0) }
             .map { layoutManager.minimumRowBounds(of: $0, in: textContainer) }
         
         let newRanges = (lineFragmentUsedRects + [newLineRect])
             .flatMap { lineRect in rowBounds
-                .filter { $0.x < lineRect.maxX }
+                .filter { ($0.x...($0.x + $0.width)).overlaps(lineRect.minX...lineRect.maxX) }
                 .map { NSRect(x: $0.x, y: lineRect.midY, width: $0.width, height: 0) }
             }
-            .map { layoutManager.glyphRange(forLineRect: $0, in: textContainer) }
-            .map { layoutManager.characterRange(forGlyphRange: $0, actualGlyphRange: nil) }
+            .map { $0.offset(by: self.textContainerOrigin) }  // to view-based
+            .map { self.lineInsertionRange(for: $0) }
         
         guard let set = self.prepareForSelectionUpdate(newRanges) else { return }
         
-        self.setSelectedRanges(set.selectedRanges, affinity: affinity, stillSelecting: false)
+        self.setSelectedRanges(set.selectedRanges, affinity: .upstream, stillSelecting: false)
         self.insertionLocations = set.insertionLocations
         self.scrollRangeToVisible(newRanges.last!)  // the last is newly added one
+    }
+    
+    
+    /// Return the range for selection that are laid out within the given rectangle
+    /// expecting the given rect is contained in a single line fragment.
+    ///
+    /// - Parameter rect: The bounding rectangle for which to return range.
+    /// - Returns: Character range corresponding to the given rectangle.
+    private func lineInsertionRange(for rect: NSRect) -> NSRange {
+        
+        let minBound = self.characterIndexForInsertion(at: NSPoint(x: rect.minX, y: rect.midY))
+        let maxBound = self.characterIndexForInsertion(at: NSPoint(x: rect.maxX, y: rect.midY))
+        
+        return NSRange(min(minBound, maxBound)..<max(minBound, maxBound))
     }
     
 }
 
 
 
-@objc extension NSTextView {
+extension NSTextView {
     
     /// Calculate rect for insartion point at `index`.
     ///
     /// - Parameter index: The character index where the insertion point will locate.
     /// - Returns: Rect where insertion point filled.
-    func insertionPointRect(at index: Int) -> NSRect {
+    @objc func insertionPointRect(at index: Int) -> NSRect {
         
         guard
             let layoutManager = self.layoutManager,
@@ -362,11 +391,7 @@ extension MultiCursorEditing where Self: NSTextView {
         
         return NSRect(x: floor(rect.minX), y: rect.minY, width: 1 / self.scale, height: rect.height)
     }
-}
-
-
-
-extension NSTextView {
+    
     
     /// Find the location for a insertion point where one (visual) line above to the given insertion point location.
     ///
@@ -471,52 +496,57 @@ private extension NSLayoutManager {
     ///            when the range extends across multiple lines.
     func minimumRowBounds(of glyphRange: NSRange, in container: NSTextContainer) -> (x: CGFloat, width: CGFloat) {
         
-        let lowerRect = self.boundingRect(forGlyphRange: NSRange(location: glyphRange.lowerBound, length: 0), in: container)
-        let upperRect = self.boundingRect(forGlyphRange: NSRange(location: glyphRange.upperBound, length: 0), in: container)
+        let lowerX = self.boundingRect(forGlyphRange: NSRange(location: glyphRange.lowerBound, length: 0), in: container).minX
+        let upperX = self.boundingRect(forGlyphRange: NSRange(location: glyphRange.upperBound, length: 0), in: container).minX
         
-        return (x: min(lowerRect.minX, upperRect.minX), width: abs(lowerRect.minX - upperRect.minX))
-    }
-    
-    
-    /// Return the range for glyphs that are laid out within the given rectangle in the given text container
-    /// expecting the given rect is contained in a single line fragment.
-    ///
-    /// - Parameters:
-    ///   - rect: The bounding rectangle for which to return glyphs.
-    ///   - textContainer: The container in which the returned glyph is laid out.
-    /// - Returns: Glyph range corresponding to the given rectangle.
-    /// - Note: Not like formal `glyphRange(forBoundingRect:in:)`, this method returns non-zero location
-    ///         even when the given rect is empty.
-    func glyphRange(forLineRect rect: NSRect, in container: NSTextContainer) -> NSRange {
-        
-        let lowerGlyphIndex = self.glyphIndex(for: NSPoint(x: rect.minX, y: rect.minY), in: container)
-        let upperGlyphIndex = self.glyphIndex(for: NSPoint(x: rect.maxX, y: rect.minY), in: container)
-        
-        return NSRange(lowerGlyphIndex..<upperGlyphIndex)
+        return (x: min(lowerX, upperX), width: abs(lowerX - upperX))
     }
     
     
     /// Return all line fragment used rects including `extraLineFragmentUsedRect` or empty range at the end of given range.
     ///
-    /// - Parameter characterRange: The glyph range where to return line fragment rectangles.
+    /// - Parameters:
+    ///   - glyphRange: The glyph range where to return line fragment rectangles.
+    ///   - effectiveRange: On output, the range for all glyphs in the line fragments.
     /// - Returns: An array of the portions of the line fragment rectangles that actually contains glyphs or other marks that are drawn.
-    func lineFragmentUsedRects(in glyphRange: NSRange) -> [NSRect] {
+    func lineFragmentUsedRects(inSelectedGlyphRanges glyphRanges: [NSRange], effectiveRange: inout NSRange) -> [NSRect] {
+        
+        assert(!glyphRanges.isEmpty)
         
         var rects: [NSRect] = []
-        var glyphIndex = glyphRange.lowerBound
-        while glyphIndex <= glyphRange.upperBound {
-            var effectiveRange: NSRange = .notFound
-            let rect = self.lineFragmentUsedRect(forGlyphAt: glyphIndex, effectiveRange: &effectiveRange, withoutAdditionalLayout: true)
-            glyphIndex = effectiveRange.upperBound
-            
-            rects.append(rect)
-        }
+        effectiveRange = glyphRanges.first!
         
-        if glyphRange.upperBound == self.numberOfGlyphs {
-            rects.append(self.extraLineFragmentUsedRect)
+        for glyphRange in glyphRanges {
+            if glyphRange.length > 0 {
+                var localEffectiveRange = glyphRange
+                self.enumerateLineFragments(forGlyphRange: glyphRange) { (_, usedRect, _, effectiveLineRange, _) in
+                    rects.append(usedRect)
+                    localEffectiveRange.formUnion(effectiveLineRange)
+                }
+                effectiveRange.formUnion(localEffectiveRange)
+                
+            } else {
+                if glyphRange.location != self.numberOfGlyphs || self.extraLineFragmentTextContainer == nil {
+                    // -> clamp the bound with `numberOfGlyphs - 1`
+                    //    Because passing `numberOfGlyphs` to `lineFragmentUsedRect(forGlyphAt:effectiveRange:)` is invalid
+                    //    and causes the warning: `_NSLayoutTreeLineFragmentUsedRectForGlyphAtIndex`. (2019-02)
+                    let safeGlyphIndex = min(glyphRange.location, self.numberOfGlyphs - 1)
+                    
+                    var effectiveLineRange: NSRange = .notFound
+                    let usedRect = self.lineFragmentUsedRect(forGlyphAt: safeGlyphIndex, effectiveRange: &effectiveLineRange, withoutAdditionalLayout: true)
+                    
+                    rects.append(usedRect)
+                    effectiveRange.formUnion(effectiveLineRange)
+                    
+                } else {
+                    rects.append(self.extraLineFragmentUsedRect)
+                    effectiveRange.formUnion(glyphRange)
+                }
+            }
         }
+        assert(!rects.isEmpty)
         
-        return rects
+        return rects.unique
     }
     
 }