+++ Tests/StringIndentationTests.swift
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2015-2018 1024jp
+//  © 2015-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -79,14 +79,16 @@ class StringIndentationTests: XCTestCase {
     
     func testSoftTabDeletion() {
         
-        let string = "     foo\n  bar"
+        let string = "     foo\n  bar   "
         
         XCTAssertNil(string.rangeForSoftTabDeletion(in: NSRange(0..<0), tabWidth: 2))
         XCTAssertNil(string.rangeForSoftTabDeletion(in: NSRange(4..<5), tabWidth: 2))
         XCTAssertNil(string.rangeForSoftTabDeletion(in: NSRange(6..<6), tabWidth: 2))
         XCTAssertEqual(string.rangeForSoftTabDeletion(in: NSRange(5..<5), tabWidth: 2)!, NSRange(4..<5))
         XCTAssertEqual(string.rangeForSoftTabDeletion(in: NSRange(4..<4), tabWidth: 2)!, NSRange(2..<4))
-        XCTAssertEqual(string.rangeForSoftTabDeletion(in: NSRange(10..<10), tabWidth: 2)!, NSRange(9..<10))
+        XCTAssertNil(string.rangeForSoftTabDeletion(in: NSRange(10..<10), tabWidth: 2))
+        XCTAssertEqual(string.rangeForSoftTabDeletion(in: NSRange(11..<11), tabWidth: 2)!, NSRange(9..<11))
+        XCTAssertNil(string.rangeForSoftTabDeletion(in: NSRange(16..<16), tabWidth: 2))
     }
 
 }